move apps to src, use it to populate applications list
